Added tray Icon, context menu, shortcut disable with memory
Whenever you try to close the app, it minimizes to the tray Icon (Notify icon) on the taskbar.
You can close the app by right clicking on the Notify Icon.
Changed the shortcut from F3 to F12.
Option to disable the shortcut.
It memorizes your settings, if you close the program with shortcut disabled, it will remember the last time that you had it disabled, using Properties.Settings. Check the code for more info.